+++ packages/discord.js/src/structures/ModalSubmitInteraction.js
@@ -1,10 +1,10 @@
 'use strict';
 
+const { lazy } = require('@discordjs/util');
 const BaseInteraction = require('./BaseInteraction');
 const InteractionWebhook = require('./InteractionWebhook');
 const ModalSubmitFields = require('./ModalSubmitFields');
 const InteractionResponses = require('./interfaces/InteractionResponses');
-const { lazy } = require('../util/Util');
 
 const getMessage = lazy(() => require('./Message').Message);
