+++ packages/proxy/src/util/util.ts
@@ -1,9 +1,5 @@
 import type { IncomingMessage, ServerResponse } from 'node:http';
-
-/**
- * Represents a potentially awaitable value
- */
-export type Awaitable<T> = PromiseLike<T> | T;
+import type { Awaitable } from '@discordjs/util';
 
 /**
  * Represents a simple HTTP request handler
