+++ packages/util/.gitignore
@@ -1,0 +1,27 @@
+# Packages
+node_modules/
+
+# Log files
+logs/
+*.log
+npm-debug.log*
+
+# Runtime data
+pids
+*.pid
+*.seed
+
+# Env
+.env
+
+# Dist
+dist/
+typings/
+
+docs/**/*
+
+# Miscellaneous
+.tmp/
+coverage/
+
+searchIndex/
