+++ packages/util/src/Equatable.ts
