+++ packages/util/src/JSONEncodable.ts
