+++ packages/util/src/functions/index.ts
@@ -1,0 +1,2 @@
+export * from './lazy.js';
+export * from './range.js';
