+++ packages/util/src/types.ts
@@ -1,0 +1,4 @@
+/**
+ * Represents a type that may or may not be a promise
+ */
+export type Awaitable<T> = PromiseLike<T> | T;
