+++ packages/util/tsup.config.js
@@ -1,0 +1,5 @@
+import { createTsupConfig } from '../../tsup.config.js';
+
+export default createTsupConfig({
+	minify: true,
+});
