+++ packages/ws/src/index.ts
@@ -8,7 +8,6 @@ export * from './strategies/sharding/WorkerShardingStrategy.js';
 
 export * from './utils/constants.js';
 export * from './utils/IdentifyThrottler.js';
-export * from './utils/utils.js';
 
 export * from './ws/WebSocketManager.js';
 export * from './ws/WebSocketShard.js';
