+++ packages/ws/src/utils/constants.ts
@@ -2,9 +2,9 @@ import { readFileSync } from 'node:fs';
 import { join } from 'node:path';
 import process from 'node:process';
 import { Collection } from '@discordjs/collection';
+import { lazy } from '@discordjs/util';
 import { APIVersion, GatewayOpcodes } from 'discord-api-types/v10';
 import type { OptionalWebSocketManagerOptions, SessionInfo } from '../ws/WebSocketManager.js';
-import { lazy } from './utils.js';
 
 /**
  * Valid encoding types
